perf(fast-td3): compile full update steps and reduce sync overhead
- throttle learner statistics reporting
- publish policy revisions using int64 env steps
- reject unsupported async, batched, and RNN modes
- cover warmup, resume, parameter sync, and CUDA compilation

+++ sample_factory/algo/utils/shared_buffers.py
@@ -239,6 +239,6 @@ class BufferMgr(Configurable):
                 for i in range(num_buffers):
                     self.traj_buffer_queues[device].put(i)
 
-        self.policy_versions = torch.zeros([cfg.num_policies], dtype=torch.int32)
+        self.policy_versions = torch.zeros([cfg.num_policies], dtype=torch.int64)
         if share:
             self.policy_versions.share_memory_()
